requirements.txt
# Core
langchain>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0

# LLM Providers
langchain-google-genai>=1.0.0   # Google Gemini
langchain-openai>=0.1.0          # OpenAI, Perplexity, DeepSeek, Grok, Groq, Together, Fireworks
langchain-anthropic>=0.1.0       # Anthropic Claude
langchain-mistralai>=0.1.0       # Mistral AI
langchain-cohere>=0.1.0          # Cohere
langchain-ollama>=0.1.0          # Ollama (local)

# Analysis
pandas>=2.0.0
matplotlib>=3.7.0
